Before the tests and the code quality was lacking. Now the tests and the code quality checker are passing.
From now on, in the platforms is needed to specify
all the possible prompts. This is a better way
of managing passing the prompts and as well is
needed for the tests.

It is needed for the tests as in .yaml files simple
brackets were not working properly. In the tests,
there are several tests under test_platforms.py
that are testing the prompts and for errors in the
platforms files.

Please, check the tests and the platforms files
for the correct format always. This is needed in order
to ensure a good codebase.