Fixed small bugs in the tests (because of change in the file platforms.en.md). As well, it has been updated the library to Python 3.12 (now it works).
The previous error was due to the fact that
we were checking the library without closing
the threads. From Python3.12 onwards, there is
a test that checks if the threads are closed
correctly. Before we were not closing the threads
with a code like:

```python
net = FakeNOS(inventory="inventory.yaml")
net.start()
```

If we want to avoid this error, we must close the
threads with the following code:
```python
net = FakeNOS(inventory="inventory.yaml")
net.start()
net.stop() # THIS WAS MISSING
```

This solves the issue https://github.com/fakenos/fakenos/issues/11