This commit solves the issue #16.
Basically the main problem was that at runtime
the plugins Python modules were not being loaded
correctly. They were creating a new platform with
the name FakeNOS (default) which is completely false.
What's more, in the case they were loaded correctly
the platform overrided entirely the other older platform
instead of only being selective about the commands
being overrided. Now this is issue is solved and at
runtime the plugins are loaded correctly. First it loads
the YAML and then the Python modules. If the Python module
collides with the YAML it will override only that
specific command.

The tests were also updated to reflect this change.