Feature hot reload completed.
We are building a testing framework for network equipment,
and therefore, as with any testing platform, it is important
to ensure that it is easy to deploy on. In this case, I have
implemented a hot-reloader to ensure an smooth development process
whenever we are developing new platforms.

Specifically, in this case it only reloads the commands but not
the entire connection, avoiding the need to establish the
connection via ssh from scrach. So in case we change a command like
`show clock`, we can do the change and only the output of `show clock`
will be changed. At the same time, if we add a new command we can see
it as well.

One key thing I've learned is that easiness is the key to a
successful project. Let's continue!